+++ test/dummy/config/initializers/recording_studio_support.rb
@@ -2,8 +2,12 @@
 
 RecordingStudioSupport.configure do |config|
   config.pages_path = "/support"
-  config.pages_title = "Help"
-  config.pages_subtitle = "Answers you can share."
-  config.admin_section_title = "Help"
-  config.admin_section_subtitle = "Pages people use when they get stuck."
+  config.public_pages_path = "/help"
+  config.help_title = "Help"
+  config.help_subtitle = "Answers you can share."
+  config.public_help_title = "Help"
+  config.public_help_subtitle = "Answers you can read."
+  config.admin_help_title = "Help"
+  config.admin_help_subtitle = "Pages people use when they get stuck."
 end
+